move GuessResult to common
