+++ battleship-common/src/lib.rs
@@ -13,3 +13,24 @@ pub mod board {
 
 pub use board::BoardView;
 
+/// Result of a guess on the game board.
+#[derive(Debug, PartialEq)]
+pub enum GuessResult {
+    /// Shot missed all ships.
+    Miss,
+    /// Shot hit a ship but didn't sink it.
+    Hit,
+    /// Shot hit and sunk a ship (includes ship name).
+    Sunk(&'static str),
+}
+
+impl std::fmt::Display for GuessResult {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            GuessResult::Miss => write!(f, "Miss"),
+            GuessResult::Hit => write!(f, "Hit"),
+            GuessResult::Sunk(name) => write!(f, "The {} was sunk!", name),
+        }
+    }
+}
+
