+++ battleship-core/src/fleet.rs
@@ -2,12 +2,12 @@ use std::collections::HashSet;
 
 use crate::constants::GameplayError;
 use crate::constants::GuessError;
-use crate::constants::GuessResult;
+use battleship_common::GuessResult;
 
 use crate::constants::SHIPS;
 
 use crate::constants::GameplayError::ShipNotFound;
-use crate::constants::GuessResult::{Hit, Miss, Sunk};
+use battleship_common::GuessResult::{Hit, Miss, Sunk};
 use crate::ship::Ship;
 
 use array_init::array_init;
