+++ battleship-core/src/lib.rs
@@ -7,4 +7,5 @@ pub use board::Board;
 pub use battleship_common::BoardView;
 pub use fleet::Fleet;
 pub use ship::Ship;
-pub use constants::{GRID_SIZE, PlayerState, GuessResult, GuessError, GameplayError, Cell};
+pub use constants::{GRID_SIZE, PlayerState, GuessError, GameplayError, Cell};
+pub use battleship_common::GuessResult;
